Move mapwfs2 inside mapmodule and combine localizations
