+++ bundles/framework/mapfull/automagicPlugins.js
@@ -2,6 +2,7 @@
  * Map plugins that will be started even without bundle plugin configuration.
  */
 export const automagicPlugins = [
+    'Oskari.wfsvector.WfsVectorLayerPlugin',
     'Oskari.mapframework.mapmodule.VectorTileLayerPlugin',
     'Oskari.mapframework.mapmodule.BingMapsLayerPlugin',
     'Oskari.mapframework.bundle.mapmodule.plugin.PinchZoomResetPlugin'
