+++ bundles/mapping/mapmodule/AbstractMapModule.js
@@ -41,6 +41,7 @@ import './plugin/wmslayer/WmsLayerPlugin.ol';
 import './plugin/vectorlayer/VectorLayerPlugin.ol';
 import './plugin/vectortilelayer/VectorTileLayerPlugin';
 import './plugin/bingmapslayer/BingMapsLayerPlugin';
+import './plugin/wfsvectorlayer/WfsVectorLayerPlugin.ol';
 
 import './event/MapClickedEvent';
 import './event/MapMoveStartEvent';
