+++ bundles/mapping/mapmodule/event/WFSFeaturesSelectedEvent.js
