+++ bundles/mapping/mapmodule/request/ActivateHighlightRequest.js
