--- bundles/mapping/mapwfs2/resources/locale/en.js
@@ -1,11 +1,0 @@
-Oskari.registerLocalization({
-    "lang": "en",
-    "key": "MapWfs2",
-    "value": {
-        "editLayer": "Edit map layer",
-        "layerFilter": {
-            "tooltip": "Show only vector layers",
-            "featuredata": "Vector layers"
-        }
-    }
-});
