--- bundles/mapping/mapwfs2/resources/locale/fi.js
@@ -1,11 +1,0 @@
-Oskari.registerLocalization({
-    "lang": "fi",
-    "key": "MapWfs2",
-    "value": {
-        "editLayer": "Muokkaa tasoa",
-        "layerFilter": {
-            "tooltip": "Näytä vain vektoritasot",
-            "featuredata": "Vektoritasot"
-        }
-    }
-});
