--- bundles/mapping/mapwfs2/resources/locale/fr.js
@@ -1,10 +1,0 @@
-Oskari.registerLocalization({
-    "lang": "fr",
-    "key": "MapWfs2",
-    "value": {
-        "layerFilter": {
-            "tooltip": "Afficher uniquement les couches vectorielles",
-            "featuredata": "Couches vectorielles"
-        }
-    }
-});
