--- bundles/mapping/mapwfs2/resources/locale/ru.js
@@ -1,10 +1,0 @@
-Oskari.registerLocalization({
-    "lang": "ru",
-    "key": "MapWfs2",
-    "value": {
-        "layerFilter": {
-            "tooltip": "Показывать только векторные слои",
-            "featuredata": "Векторные слои"
-        }
-    }
-});
