--- bundles/mapping/mapwfs2/resources/locale/sv.js
@@ -1,11 +1,0 @@
-Oskari.registerLocalization({
-    "lang": "sv",
-    "key": "MapWfs2",
-    "value": {
-        "editLayer": "Editera kartlager",
-        "layerFilter": {
-            "tooltip": "Visa endast kartlager med vektorgrafik",
-            "featuredata": "Vektorlager"
-        }
-    }
-});
